# contract/openapi.yaml
openapi: 3.1.0
info:
  title: КОСМОБЕРЕГ API
  version: 1.0.0
  description: |
    Контракт между фронтендом и бэкендом платформы КОСМОБЕРЕГ.

    ## Как этим пользоваться

    Это **единственный источник истины** об интерфейсе между фронтом и бэком.
    Фронт кодируется против контракта (с моками, пока бэка нет), бэк реализует
    контракт. Стыковка = смена базового URL.

    Изменение контракта — только по согласованию обеих сторон. Правило простое:
    **сначала правим contract/openapi.yaml, потом код.**

    ## Ключевое решение: локализация на клиенте

    API **никогда не возвращает переведённый текст**. Возвращаются доменные коды
    (`status: "problem"`, `role: "observer"`) и ключи локализации
    (`name_key: "seg.ustie.name"`). Клиент переводит их через `i18n.js`.

    Причины:
    - фронт уже несёт словарь на 3 языка (ru/en/zh) — дублировать его на бэке не нужно;
    - смена языка не требует запроса к серверу;
    - добавление 4-го языка не затрагивает бэкенд вообще.

    Исключение — **пользовательский контент** (описания в отчётах, имена
    пользователей): возвращается как есть, без перевода.

    ## Что вычисляется только на сервере

    Клиент не может это подделать, поэтому такие значения приходят с бэка и
    никогда не принимаются от фронта:

    - `attention_index` и факторы `s`, `c`, `t`, `a`;
    - переход роли `student → observer` (только после завершения курса);
    - достижение консенсуса по разметке;
...
    - смена статуса участка.
servers:
...
security:
  - bearerAuth: []
paths:
...
  /events:
    get:
      tags: [events]
      summary: Список акций
      parameters:
        - name: segment_id
          in: query
          schema: { type: string }
        - name: upcoming_only
          in: query
          schema: { type: boolean, default: true }
      responses:
        '200':
          description: Акции
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Event' }
    post:
      tags: [events]
      summary: Создать акцию
      description: |
        Только роль `oopt_staff`. Волонтёр не может назначить выход на охраняемую
        территорию — это требование режима ООПТ. Роль `ambassador` может подать
        заявку, но она требует подтверждения сотрудником.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [segment_id, starts_at, capacity]
              properties:
                segment_id: { type: string }
                starts_at:  { type: string, format: date-time }
                capacity:   { type: integer, minimum: 1, maximum: 500, example: 25 }
                meeting_point: { type: string, maxLength: 300 }
      responses:
        '201':
          description: Акция создана
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Event' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422':
          description: Участок не подтверждён — акцию создать нельзя
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /events/{event_id}/enrollment:
    post:
      tags: [events]
      summary: Записаться на акцию
      description: |
        Сервер возвращает список **невыполненных условий допуска** в
        `blocking_requirements`. Фронт по нему решает, какой экран показать
        дальше: согласие родителя или инструктаж.
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '201':
          description: Запись создана
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Enrollment' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Мест нет или уже записан
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /events/{event_id}/consent:
    post:
      tags: [events]
      summary: Запросить согласие законного представителя
      description: |
        Только для `is_minor: true`. Сервер отправляет ссылку на подписание.
        До подтверждения допуск закрыт — проверка выполняется на сервере
        и не может быть обойдена с клиента.
      parameters:
        - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [parent_contact]
              properties:
                parent_contact:
                  type: string
                  description: Email или телефон родителя
      responses:
        '202':
          description: Запрос отправлен, ожидается подтверждение
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Enrollment' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /events/{event_id}/briefing:
    post:
      tags: [events]
      summary: Подтвердить прохождение инструктажа
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Инструктаж засчитан
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Enrollment' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /events/{event_id}/consent/{token}/sign:
    post:
      tags: [events]
      security: []
      summary: Подписание согласия законным представителем
      description: |
        Открывается по одноразовой ссылке из письма родителю. Без авторизации:
        у родителя нет аккаунта, и заводить его ради одного действия означало
        бы собирать лишние персональные данные. Защита — неугадываемый токен
        со сроком жизни 14 дней.

        Экран подписания — публичная страница, отдельная от приложения.
      parameters:
        - $ref: '#/components/parameters/EventId'
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Согласие подписано, допуск пересчитан
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Enrollment' }
        '404':
          description: Ссылка не найдена
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Согласие уже подписано
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          description: Срок действия ссылки истёк
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ==================== S6: полевой отчёт ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    EventId:
      name: event_id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    Forbidden:
      description: Роль не позволяет выполнить операцию
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [code, message_key]
      properties:
        code:
          type: string
          example: consent_required
          description: Машинный код ошибки
        message_key:
          type: string
          example: err.consent_required
          description: Ключ локализации для i18n.js — текст берётся из словаря фронта
        details:
          type: object
          additionalProperties: true
...
    Event:
      type: object
      required: [id, segment_id, starts_at, capacity, enrolled_count, status]
      properties:
        id:             { type: string, format: uuid }
        segment_id:     { type: string }
        segment_name_key: { type: string }
        starts_at:      { type: string, format: date-time }
        capacity:       { type: integer }
        enrolled_count: { type: integer }
        meeting_point:  { type: string }
        status:         { type: string, enum: [planned, active, finished, cancelled] }
        organizer_oopt_id: { type: string }
        is_my_segment:
          type: boolean
          description: |
            true, если текущий пользователь сам размечал этот участок.
            Фронт использует это для механики персональной привязки —
            ключевого перехода «онлайн → поле».
    Enrollment:
      type: object
      required: [event_id, status, blocking_requirements]
      properties:
        event_id: { type: string, format: uuid }
        status:
          type: string
          enum: [pending_requirements, ready, attended, cancelled]
        blocking_requirements:
          type: array
          description: Что осталось выполнить для допуска — фронт ведёт пользователя по этому списку
          items:
            type: string
            enum: [course_completion, parent_consent, briefing]
        consent_status:
          type: string
          nullable: true
          enum: [not_required, requested, signed]
        briefing_completed_at: { type: string, format: date-time, nullable: true }
...